--- packages/server/db_connect.ts
@@ -1,27 +1,0 @@
-import { Sequelize, SequelizeOptions } from 'sequelize-typescript'
-import { formRecordModel } from './models/ForumRecord'
-const { POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, POSTGRES_PORT } =
-  process.env
-
-const sequelizeOptions: SequelizeOptions = {
-  host: 'localhost',
-  port: Number(POSTGRES_PORT),
-  username: POSTGRES_USER,
-  password: POSTGRES_PASSWORD,
-  database: POSTGRES_DB,
-  dialect: 'postgres',
-}
-
-export const sequelize = new Sequelize(sequelizeOptions)
-
-export const ForumRecord = sequelize.define('ForumRecord', formRecordModel, {})
-
-export const dbConnect = async () => {
-  try {
-    await sequelize.authenticate()
-    await sequelize.sync()
-    console.log('Connection has been established successfully.')
-  } catch (error) {
-    console.error('Unable to connect to the database:', error)
-  }
-}
